+++ web-client/src/components/SrMap.vue
@@ -39,8 +39,7 @@
   import { onActivated } from "vue";
   import { onDeactivated } from "vue";
   import SrCurrentMapViewParms from "./SrCurrentMapViewParms.vue";
-  import { Deck } from '@deck.gl/core/typed';
-  import { toLonLat} from 'ol/proj';
+  import { updateDeck } from "@/composables/SrMapUtils";
 
   const srToastStore = useSrToastStore();
 
@@ -85,49 +84,7 @@
     }
   };
 
-  
-  // Custom Render Logic: The render option is a function that takes an object containing size and viewState. This function is where you align the DeckGL layer's view with the OpenLayers map's current view state.
-  // size is an array [width, height] indicating the dimensions of the map's viewport.
-  // viewState contains the current state of the map's view, including center coordinates, zoom level, and rotation. This information is converted and passed to DeckGL to ensure both visualizations are synchronized.
-  // Setting DeckGL Properties: Inside the render function, properties of the DeckGL instance (deck) are updated to match the current size and view state of the OpenLayers map. This ensures that the DeckGL visualization aligns correctly with the map's viewport, zoom level, and rotation.
-  // Redrawing DeckGL: After updating the properties, deck.redraw() is called to render the DeckGL layer with the new settings.
-  // Sync deck view with OL view
-
-  function createDeckGLInstance(tgt:HTMLDivElement): Layer | null{
-      try{
-          const deck = new Deck({
-              initialViewState: {longitude: 0, latitude: 0, zoom: 1},
-              controller: false,
-              parent: tgt,
-              style: {pointerEvents: 'none', zIndex: '1'},
-              layers: []
-          });
-          mapStore.setDeckInstance(deck);
-          const layerOptions = {
-              title: 'DeckGL Layer',
-          }
-          const deckLayer = new Layer({
-              render: ({size, viewState}: {size: number[], viewState: {center: number[], zoom: number, rotation: number}})=>{
-                  const [width, height] = size;
-                  const [longitude, latitude] = toLonLat(viewState.center);
-                  const zoom = viewState.zoom - 1;
-                  const bearing = (-viewState.rotation * 180) / Math.PI;
-                  const deckViewState = {bearing, longitude, latitude, zoom};
-                  deck.setProps({width, height, viewState: deckViewState});
-                  deck.redraw();
-                  return document.createElement('div');
-              },
-              ...layerOptions
-          });
-          return deckLayer // we just need a 'fake' Layer object with render function and title to marry to Open Layers
-      } catch (error) {
-          console.error('Error creating DeckGL instance:',error);
-          return null;
-      }
-  }
-
-
-    // Function to toggle the DragBox interaction.
+  // Function to toggle the DragBox interaction.
   function disableDragBox() {
     console.log("disableDragBox");
     // Check if the DragBox interaction is added to the map.
@@ -367,14 +324,6 @@
           map.addControl(plink);
         }
         updateMapView("onMounted");
-        const tgt = map.getViewport() as HTMLDivElement;
-        const deckLayer = createDeckGLInstance(tgt);
-        if(deckLayer){
-            map.addLayer(deckLayer);
-            console.log('deckLayer added:',deckLayer);
-        } else {
-            console.error('createDeckGLInstance returned null');
-        }
 
       } else {
         console.log("Error:map is null");
@@ -515,6 +464,7 @@
             updateCurrentParms();
             addLayersForCurrentView();      
             map.getView().on('change:resolution', onResolutionChange);
+            updateDeck(map);
             // Permalink
             // if(mapStore.plink){
             //   var url = mapStore.plink.getUrlParam('url');
