+++ web-client/src/runDevDate.ts
@@ -1,1 +1,1 @@
-export const RUN_DEV_DATE = '2024-04-24 09:41:44';
+export const RUN_DEV_DATE = '2024-04-24 11:55:23';
